requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
